requirements.txt
requests
pyyaml
psycopg2